+++ PyFlora/CHANGELOG.md
@@ -1,5 +1,9 @@
 # tracking changes and progess
 
+May 1
+
+- login GUI class
+
 Apr 25
 
 - add task - game guess the num